handle sigterm properly
This rewrites `start-l2met-shuttle`, and uses a binary from https://github.com/heroku/l2met-shuttle/pull/14 that ignores `SIGTERM`

This is because, on heroku dynos, the dyno manager sends SIGTERM to every process when doing a graceful shutdown. During that time period, we still want logging to happen and metrics to be sent.

With the prior implementation, the `tee` processes would exit the moment the dyno got told to shutdown.

In the new implementation, `l2met-shuttle` hangs around until the process it's reading from dies.